requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
line_profiler==4.2.0
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
pluggy==1.5.0
-e git+ssh://git@github.com/nicholaspearson43/devtools_scicomp_project_2025.git@a91c4d64011142e07695b698102b92838c6d0cf2#egg=pyclassify
pytest==8.3.4
PyYAML==6.0.2
tomli==2.2.1
